back-end/api/requirements.txt
langchain-groq
langchain-community
pypdf
sentence-transformers
pinecone
langchain_huggingface
langchain-qdrant
langchain-pinecone
unstructured
langgraph
